質問、回答削除を投稿者のみに制限

+++ app/src/Controller/AnswersController.php
@@ -35,7 +35,10 @@ class AnswersController extends AppController
     {
         $answer = $this->Answers->get($id);
         $questionId = $answer->question_id;
-        // @TODO 回答を削除できるのは回答投稿者のみとする
+        if ($answer->user_id != $this->Auth->user('id')) {
+            $this->Flash->error('他のユーザーの回答を削除することはできません');
+            return $this->redirect(['controller' => 'Questions', 'action' => 'view', $questionId]);
+        }
 
         if ($this->Answers->delete($answer)) {
             $this->Flash->success('回答を削除しました');
